fix(exla): do not rebuild libexla if already built
This should reduce build time when running the application in different
environments. On my machine the XLA was rebuilt in test environment,
even if it was already built in `dev` environment. This caused
unnecessary work when using ElixirLS which is building the whole project
in `test` environment. This change will simply assume that project is
built if there already is `libexla.so`.