Add fan-out to grads (#564)
Our grad was incorrect when the same expression appeared
more than once. To address it, we need to collect parents (all
nodes that receive a node as argument) and add their g together.
You can find a discussion of how this should be solved conceptually
in this article:

http://h2.jaguarpaw.co.uk/posts/automatic-differentiation-worked-examples/

In practice, we traverse the nodes, building the child => parent
graphs, and then we make it so the parents fill in the gs of the
children, until all parents are traversed recursively.

We have removed outer, as it can be implemented using existing
primitives, and product was temporarily removed for now.