Implement Nx.LinAlg.matrix_power/2 (#582)
This PR implements the matrix_power function as requested in #582,
using repeated squaring to calculate the power in logarithmic
time.

Two doctests are provided, one to make sure that identity matrices
don't change, and one to test a simple exponent case.
This should be enoguh to catch serious implementation issues, and
enough to provide documentation on what this function actually does.

Note that technically, if we were to ever support arbitrarily large
numbers, this algorithm would take more in the neighborhood of
linear time, as the multiplication of very large integers becomes
slower with the width of such integers.