Unify defn arguments
Prior to this patch, `compile`/`jit` accepted only
tensors while `defn` allowed functions and keyword
lists as default arguments.

This PR unifies the approaches such that functions
and lists are allowed in all APIs, regardless if a
default argument or not.

This simplifies the implementation and makes the
process of compiling/jitting a `defn` simpler.