feat: send telemetry event after XLA compilation
This commit adds telemetry events that are executed when EXLA compiles a
function into an executable.

These events can be used to have observability into how many times
functions are compiled and how long compiling them takes.

When a function is compiled, an event with name `[:exla, :compilation]`
is executed. The measurements include the following:
- `eval_time` - time it took to call the defn function that is compiled
- `compile_time` - time it took for XLA to compile the instructions into
an executable
- `total_time` - sum of `eval_time` and `compile_time`

This commit is related to the issue
https://github.com/elixir-nx/nx/issues/974