Output of complex/2 preserves the minimum precision of its inputs
The minimum precision of the inputs to Nx.complex/2 is used to
determine what precision is needed from the call to `Nx.Constants.i/1`
in order to return a complex-valued tensor with a type that matches
the minimum precision of its inputs.

This change results in a different default behavior for expressions
like `Nx.complex(Nx.tensor(1), Nx.tensor(2))`. Previously this would
return a tensor with type :c64; however, since the default type for
`Nx.tensor(1)` is :s64 complex/2 will now return a tensor with type
c128 in order to preserve the precision of the real and imaginary
parts.

Expand examples in complex/2 docs to show how this works.

Fix up types in existing doctests.

Fix types in defn_tests.ex test "complex numbers" since 64 bit inputs
now produces a tensor of type c128 instead of c64.