Use Nx.Type.merge and Nx.Type.to_complex to determin complex type
Making use of the API in Nx.Type means there is no need to our own
math to find the correct precision for the complex tensor. This fixes
the error I had introduced with respect to tensors of type :s64 and
makes the function substantially more readable.

Moved complex and distracting examples from the @doc to the test suite.