Differentiate between complex, real output in binary backend as_type
Because complex types are considered floats by `Nx.Type.float?/1` the
binary backend's implementation of as_type could not preserve the
imaginary part of a complex number when casting between :c64 and
:c128.

This addresses that error by introducing a new function
`Nx.Type.complex?/1` that is used to determine if the output of
`as_type/1` should include just the real part of the input.

I considered modifying `Nx.Type.float?/1` to return `false` for
complex types. This seems reasonable given the docs which say "is a
float in elixir" and complex is a struct. However, this change
produces a substantial number of errors in other tests. In the
interest of keeping changes to a minimum, I added the `complex?`
function and modified the binary backend's `as_type` implementation to
check whether the output is real when removing the imaginary part of a
complex input.