Require passing in contents of NumPy file instead of filename
Previously, the `Nx.from_numpy/1` and `Nx.from_numpy_archive/1`
functions took as arguments file paths.

That made it inconvenient to use this function when the NumPy
array/archive came from somewhere else. For example, if it was
downloaded from the network to memory directly.

This commit changes both `Nx.from_numpy/1` and `Nx.from_numpy_archive/1`
to take as arguments the contents of the NumPy file as binary.

I also took the liberty of adding examples to both functions to make it
clearer what they expect as inputs and what are their return values.