Return array names from `Nx.from_numpy_archive`
Previously, only arrays/tensors were returned from that function. Those
arrays also have names which were not returned.

This commit makes the function include array names in the return value.
Now, instead of returning a list of tensors, we return a list of tuples
where each tuple contains the array's name and its value.

The array names are present in the zip file as file names. In the zip
file the arrays also have a .npy extension which needs to be removed.

If there is a file with name that does not end with .npy, then the name
is kept as is. I think in practice all files in the archive should have
the .npy extension, but NumPy's own loading logic handles it like that,
and I decided to make the implementation in Elixir consistent: https://github.com/numpy/numpy/blob/b0fdcb3b80577033611d6d9ff690ecc64f1d636a/numpy/lib/npyio.py#L194

I also added typespecs to both `from_numpy` and `from_numpy_archive` to
make it clearer what is returned.