Improvements to `from_numpy` and `from_numpy_archive` (#1004)
* Require passing in contents of NumPy file instead of filename

Previously, the `Nx.from_numpy/1` and `Nx.from_numpy_archive/1`
functions took as arguments file paths.

That made it inconvenient to use this function when the NumPy
array/archive came from somewhere else. For example, if it was
downloaded from the network to memory directly.

This commit changes both `Nx.from_numpy/1` and `Nx.from_numpy_archive/1`
to take as arguments the contents of the NumPy file as binary.

I also took the liberty of adding examples to both functions to make it
clearer what they expect as inputs and what are their return values.

* Return array names from `Nx.from_numpy_archive`

Previously, only arrays/tensors were returned from that function. Those
arrays also have names which were not returned.

This commit makes the function include array names in the return value.
Now, instead of returning a list of tensors, we return a list of tuples
where each tuple contains the array's name and its value.

The array names are present in the zip file as file names. In the zip
file the arrays also have a .npy extension which needs to be removed.

If there is a file with name that does not end with .npy, then the name
is kept as is. I think in practice all files in the archive should have
the .npy extension, but NumPy's own loading logic handles it like that,
and I decided to make the implementation in Elixir consistent: https://github.com/numpy/numpy/blob/b0fdcb3b80577033611d6d9ff690ecc64f1d636a/numpy/lib/npyio.py#L194

I also added typespecs to both `from_numpy` and `from_numpy_archive` to
make it clearer what is returned.

* Add unit tests for `Nx.from_numpy` and `Nx.from_numpy_archive`

These functions did previously not have any tests. This commit adds a
few basic test cases for them.