Make serving explicitly receive compiler options
The goal is to allow us to device partitioning
in the future that will work transparently.
For example, we could do:

    {Nx.Serving,
     serving: my_serving(),
     partitioned: true}

And, if you have two devices, it will create two
Nx.Serving and automatically partition the load
between them. This will be handy with the distributed
option too, making sure we can spread the load across
multiple GPUs and machines.